+++ src/js/module/require-config.js
@@ -5,12 +5,12 @@ require = { // eslint-disable-line no-global-assign
     'paths': {
         'vue': 'js/require/vue.js?v=2.6.12',
         'vue-router': 'js/require/vue-router.js?v=3.4.9',
-        'vue-i18n': 'js/require/vue-i18n.js?v=8.22.3',
+        'vue-i18n': 'js/require/vue-i18n.js?v=8.22.4',
         'vue-ctk-date-time-picker': 'js/require/vue-ctk-date-time-picker.js?v=2.5.0',
-        '@fortawesome/fontawesome-svg-core': 'js/require/fontawesome-svg-core.js?v=1.2.32',
-        '@fortawesome/free-brands-svg-icons': 'js/require/free-brands-svg-icons.js?v=5.15.1',
-        '@fortawesome/free-regular-svg-icons': 'js/require/free-regular-svg-icons.js?v=5.15.1',
-        '@fortawesome/free-solid-svg-icons': 'js/require/free-solid-svg-icons.js?v=5.15.1',
+        '@fortawesome/fontawesome-svg-core': 'js/require/fontawesome-svg-core.js?v=1.2.34',
+        '@fortawesome/free-brands-svg-icons': 'js/require/free-brands-svg-icons.js?v=5.15.2',
+        '@fortawesome/free-regular-svg-icons': 'js/require/free-regular-svg-icons.js?v=5.15.2',
+        '@fortawesome/free-solid-svg-icons': 'js/require/free-solid-svg-icons.js?v=5.15.2',
         '@fortawesome/vue-fontawesome': 'js/require/vue-fontawesome.js?v=2.0.2'
     }
 };